refactor: reviews
1. page 정보가 중복으로 적용된것 삭제
